TypeScript compat changes for CRA and AF
