+++ src/constants.ts
@@ -1,3 +1,5 @@
+import { Settings } from './types';
+
 export const PACKAGE_NAME = 'react-loosely-lazy';
 
 export const MODE = {
@@ -14,10 +16,6 @@ export const PHASE = {
 export const PHASE_LAZY_DELAY = 50;
 
 export const COLLECTED = new Map();
-
-export type Settings = {
-  CURRENT_MODE: typeof MODE.HYDRATE | typeof MODE.RENDER;
-};
 
 export const SETTINGS: Settings = {
   CURRENT_MODE: MODE.HYDRATE,
