+++ src/phase/index.tsx
@@ -1,5 +1,5 @@
 export { LazyWait } from './component';
-export type { LazyWaitProps } from './component';
 export { useLazyPhase, LazyPhaseContext, setCurrent } from './context';
 export { usePhaseSubscription } from './controller';
 export { LISTENERS } from './listeners';
+export { LazyWaitProps } from './types';
