+++ src/phase/types.ts
@@ -1,0 +1,6 @@
+import { ReactNode } from 'react';
+
+export type LazyWaitProps = {
+  until: boolean;
+  children: ReactNode;
+};
