+++ src/suspense/index.tsx
@@ -1,4 +1,3 @@
 export { LazySuspense } from './component';
-export type { LazySuspenseProps } from './component';
 export { LazySuspenseContext } from './context';
-export type { Fallback } from './context';
+export { Fallback, LazySuspenseContextType, LazySuspenseProps } from './types';
