+++ src/suspense/types.ts
@@ -1,0 +1,12 @@
+import { SuspenseProps } from 'react';
+
+export type Fallback = SuspenseProps['fallback'];
+
+export type LazySuspenseContextType = {
+  fallback: Fallback;
+  setFallback(fallback: Fallback): void;
+};
+
+export type LazySuspenseProps = {
+  fallback: Fallback;
+};
