+++ src/types.ts
@@ -1,0 +1,5 @@
+import { MODE } from './constants';
+
+export type Settings = {
+  CURRENT_MODE: typeof MODE.HYDRATE | typeof MODE.RENDER;
+};
